legacy/requirements.txt
streamlit
sentence-transformers
scikit-learn
pandas
numpy
google-genai
torch